requirements.txt
py-cord
pymongo
qrcode[pil]
discord_webhook
matplotlib
python-dotenv
colorama
requests
psutil
bitcoinlib
electrum-client
Pillow
web3
web3==6.0.0
eth-account==0.9.0
eth-hash[pycryptodome]
solana
solders
bip-utils